ODM: di readMatrix ganti float jadi double

+++ src/Library/OperasiDasarMatrix.java
@@ -26,7 +26,7 @@ public class OperasiDasarMatrix {
 
         for (int i = 0; i < nRow; i++) {
             for (int j = 0; j < nCol; j++) {
-                m.set_ELMT(i, j, sc.nextFloat());
+                m.set_ELMT(i, j, sc.nextDouble());
             }
         }
     }
